refactor: Input 컴포넌트 네이밍 QuantityInput으로 수정
