--- src/components/Input/Input.stories.jsx
@@ -1,12 +1,0 @@
-import Input from './Input';
-
-export default {
-  title: 'Component/Input',
-  component: Input,
-};
-
-function Template(args) {
-  return <Input {...args} />;
-}
-
-export const Default = Template.bind({});
